Fixed race between link and peer disconnection.
This fixes a rare race (segfault) between `link` and
`ignore_recv_data`.  If the peer of the socket exits between
establishing a connection and libprocess queuing a `MessageEncoder`,
`ignore_recv_data` may delete the `Socket` underneath the `link`.

This patch is meant to be easily backported.

Review: https://reviews.apache.org/r/49416/